+++ packages/local_auth/local_auth_android/android/src/test/java/io/flutter/plugins/localauth/LocalAuthTest.java
@@ -47,7 +47,7 @@ public class LocalAuthTest {
   static final AuthStrings dummyStrings =
       new AuthStrings.Builder()
           .setReason("a reason")
-          .setBiometricHint("a hint")
+          .setSignInHint("a hint")
           .setCancelButton("cancel")
           .setSignInTitle("sign in")
           .build();