+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -3,6 +3,7 @@
 * **BREAKING CHANGES:**
   * Switches to `LocalAuthException` for error reporting.
   * Removes support for `useErrorDialogs`.
+  * Renames `biometricHint` to `signInHint` to reflect its usage.
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 1.0.51